Redoing Engi/Space Allocation of Marine Control
Marine control is going to be merged with Engineering Control; and Engineering is getting a redo in general.

Engi was actually the first area I worked on in the map: I went from engineering outwards, it had a sprucening up early on when I made the weather deck concrete instead of plates but it's still really gross to me and I want to redo it.